requirements.txt
args==0.1.0
clint==0.5.1
colorama==0.4.6
filelock==3.13.1
fsspec==2023.10.0
Jinja2==3.1.2
joblib==1.3.2
jointml-client==0.2.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
pandas==2.1.3
pyfiglet==1.0.2
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.1.1
typing_extensions==4.8.0
tzdata==2023.3
